+++ CHANGELOG.md
@@ -14,6 +14,12 @@ All notable changes to this project will be documented in this file. The format 
   - Improved UX for the Building Selection Criteria drawer ([#1994](https://github.com/bloom-housing/bloom/pull/1994)) (Jared White)
   - alternate contact email is validated ([#2035](https://github.com/bloom-housing/bloom/pull/2035)) (Yazeed)
 
+## Backend
+
+- Fixed:
+
+  - Incorrect listing status ([#2015](https://github.com/bloom-housing/bloom/pull/2015)) (Dominik Barcikowski)
+
 ## v2.0.0-pre-tailwind 09/16/2021
 
 ## Frontend